Move the existing bloom filter bulk query code to its own module
The new optimised bulk query will gets its own module too. This will
make it easier to compare. It will also keep the generated code smaller
which helps when looking at generated code during optimisation.

Also, we will have to keep both versions for better GHC version
compatibility.
